Refactor: 컴포넌트 분리, Util Fucntion 분리,
